Organize resource code into a resource module
