+++ cliquet_docs/reference/storage.rst
@@ -24,31 +24,6 @@ Memory
 .. autoclass:: cliquet.storage.memory.Memory
 
 
-.. _cloud-storage:
-
-Cloud Storage
--------------
-
-.. note::
-
-    `Under construction <https://github.com/mozilla-services/kinto/pull/45>`_
-
-If the ``kinto`` package is available, it is possible to store data in
-a remote instance of *Kinto*.
-
-::
-
-    cliquet.storage_backend = kinto.storage
-    cliquet.storage_url = https://cloud-storage.services.mozilla.com
-
-See :rtd:`Kinto <kinto>` for more details.
-
-.. note::
-
-    In order to avoid double checking of OAuth tokens, the Kinto service
-    and the application can share the same cache (``cliquet.cache_url``).
-
-
 API
 ===
